Create a scatter plot using two variables in the crabs data set.

Uses a data set and two columns inputted by the used to create a scatter plot. Returns the completed scatter plot, or a message if the plot cannot be created

ScatterPlot(variable1, variable2, data)

Arguments

variable1

A column from the crabs data set inputted by the user. Will appear on the x axis of the graph.

variable2

A column from the crabs data set inputted by the user. Will appear on the y axis of the graph.

data

The chosen data set. In this case, the crabs data set

Value

plot The completed scatter plot made using the two chosen variables from the crabs data set. If a scatter plot cannot be made, a message will be displayed